+++ src/main/java/mod/tjt01/sprinkle/data/ModTags.java
@@ -1,5 +1,6 @@
 package mod.tjt01.sprinkle.data;
 
+import mod.tjt01.sprinkle.Main;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.*;
 import net.minecraft.world.entity.EntityType;
@@ -8,7 +9,13 @@ import net.minecraft.world.item.Item;
 
 public class ModTags {
     public static class Blocks {
+        public static final TagKey<Block> DOUBLE_DOOR_BLACKLIST = modTag("double_door_blacklist");
+
         public static final TagKey<Block> QUARK_VERTICAL_SLAB = quark("vertical_slab");
+
+        private static TagKey<Block> modTag(String name) {
+            return tag(new ResourceLocation(Main.MODID, name));
+        }
 
         private static TagKey<Block> quark(String name) {
             return tag(new ResourceLocation("quark", name));